Initial implementation of structured log support
Added implementation of a LoggerContext which provides support for
structured logs. Along with the implementation of the LoggerContext it
has been added to the Logger struct a field to receive a formatter
that will be used when the log entry is being sent to the configured
output.

When using the LoggerContext instead of the Logger, the message passed
in the logging methods (Info, Warning, Error, Debug...) will be included in
the as part of the JSON object in the key "msg" together with all the other
fields specified when using the WithContext method.