Write fields when using TextFormatter
To improve readability of the log while in development, it would be good to
have the logs in text format.

Changes:
- Skip setting the JSONFormatter by default when calling SetGlobalFields and
WithScope methods, and use the formatter that is set in the global logger
instance instead.

- Introduced a private function called writeFields which is called when
using the Textformatter and it will print in plain text format the keys
and values of the Fields map.